+++ spec/grape/validations/validators/at_least_one_of_validator_spec.rb
@@ -244,5 +244,33 @@ describe Grape::Validations::Validators::AtLeastOneOfValidator do
         end
       end
     end
+
+    describe 'string group attrs under build_with :hash' do
+      let(:app) do
+        Class.new(Grape::API) do
+          build_with :hash
+
+          rescue_from Grape::Exceptions::ValidationErrors do |e|
+            error!(e.errors.transform_keys! { |key| key.join(',') }, 400)
+          end
+
+          params do
+            optional :beer
+            optional :wine
+            at_least_one_of 'beer', 'wine'
+          end
+          post do
+          end
+        end
+      end
+
+      let(:path) { '/' }
+      let(:params) { { beer: true } }
+
+      it 'does not return a validation error' do
+        validate
+        expect(last_response.status).to eq 201
+      end
+    end
   end
 end